+++ libntp/strl_obsd.c
@@ -46,7 +46,7 @@ strlcpy(char *dst, const char *src, size_t siz)
 			;
 	}
 
-	return(s - src - 1);	/* count does not include NUL */
+	return((size_t)(s - src - 1));	/* count does not include NUL */
 }
 #endif				/* + */
 
@@ -73,7 +73,7 @@ strlcat(char *dst, const char *src, size_t siz)
 	/* Find the end of dst and adjust bytes left but don't go past end */
 	while (n-- != 0 && *d != '\0')
 		d++;
-	dlen = d - dst;
+	dlen = (size_t)(d - dst);
 	n = siz - dlen;
 
 	if (n == 0)
@@ -87,6 +87,6 @@ strlcat(char *dst, const char *src, size_t siz)
 	}
 	*d = '\0';
 
-	return(dlen + (s - src));	/* count does not include NUL */
+	return(dlen + (size_t)(s - src));	/* count does not include NUL */
 }
 #endif				/* + */
